[FE] refactor : 회고 폼 생성 시 유효성 검사 분리 및 상수 추가
