+++ frontend/src/service/@shared/constants/index.ts
@@ -68,7 +68,11 @@ const MYPAGE_TAB = {
   MY_REVIEW_FORMS: 'my-review-forms',
 };
 
+const REVIEW_FORM_TITLE_LENGTH = 100;
+
 const REVIEW_FORM_CODE_LENGTH = 8;
+
+const REVIEW_FORM_CODE_REGEX = `(?=.*[A-Za-z])(?=.*[0-9])[a-zA-Z0-9]{${REVIEW_FORM_CODE_LENGTH}}$`;
 
 export {
   PAGE_LIST,
@@ -82,5 +86,7 @@ export {
   ACCESS_TOKEN_REFRESH_TIME,
   MYPAGE_TAB,
   PERMISSION_VALID_TIME,
+  REVIEW_FORM_TITLE_LENGTH,
   REVIEW_FORM_CODE_LENGTH,
+  REVIEW_FORM_CODE_REGEX,
 };
